feat: 가격 타입 Long -> Decimal로 변경

+++ src/main/java/depth/finvibe/investment/modules/market/domain/CurrentPrice.java
@@ -1,9 +1,10 @@
 package depth.finvibe.investment.modules.market.domain;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.Objects;
 
-public record CurrentPrice(Long stockId, LocalDateTime at, Long price, Long open, Long high, Long low, Long close,
+public record CurrentPrice(Long stockId, LocalDateTime at, BigDecimal price, BigDecimal open, BigDecimal high, BigDecimal low, BigDecimal close,
                            float prevDayChangePct, Long volume, Long value) {
 
     @Override
